I reorganized the package structure to include the imaging_pipeline. Added a test for the make_output_dict function. For ticket #81.
