+++ mtpipeline/imaging/cosmics.py
