+++ mtpipeline/imaging/iter_cosmics.py
