+++ mtpipeline/imaging/run_astrodrizzle.py
