+++ mtpipeline/imaging/run_cosmics_test.py
