+++ mtpipeline/imaging/run_trim.py
