+++ mtpipeline/imaging/run_trim_test.py
